requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
mediapipe>=0.10.0
Pillow>=9.0.0